Refactor Protocol base class to core module
