+++ server/core/__init__.py
@@ -1,9 +1,11 @@
 from .dependency_injector import DependencyInjector
+from .protocol import Protocol
 from .routing import RouteError, Router
 from .service import Service, create_services
 
 __all__ = (
     "DependencyInjector",
+    "Protocol",
     "RouteError",
     "Router",
     "Service",
